SecretKey is now set from config for KRaft.
