+++ core/src/main/scala/kafka/zk/KafkaZkClient.scala
@@ -1527,7 +1527,7 @@ class KafkaZkClient private[zk] (zooKeeperClient: ZooKeeperClient, isSecure: Boo
       retryRequestUntilConnected(createRequest)
     }
 
-    val tokenInfo = DelegationTokenInfoZNode.encode(token)
+    val tokenInfo = DelegationTokenInfoZNode.encode(token.tokenInfo())
     val setDataResponse = set(tokenInfo)
     setDataResponse.resultCode match {
       case Code.NONODE =>
